requirement.txt
certifi==2021.10.8
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
wincertstore==0.2
